`forc-pkg`: Refactor `BuildPlan` construction and validation to minimize fetching, improve offline support, reduce I/O, fix bugs. (#2234)
* BuildPlan generation from lock file does not break when there is a dependency removal from manifest file.

* While removing the deleted nodes handle orphaned childs

* removed debug println

* removed_deps removed from BuildPlan

* While checking from parent's manifest check for package name as well

* removed unnecessary reference

* split validate into generate_diff and validate

* apply_pkg_diff visits nodes to delete in reverse BFS order

* docs added for generate_diff

* removed unnecessary logic in apply_diff_after_lock

* doc updates

* clippy && extra dereferencing removed

* Apply suggestions from code review effor -> effort

Co-authored-by: mitchmindtree <mitchell.nordine@fuel.sh>

* duplicate code removed

* change visibility of apply_pkg_diff and from_lock to private

* doc updated

* patch checks are added

* check for patches corrected

* lsp uses load_from_manifest, from_lock_file is private

* add use::collections::HashSet

* WIP Improve `BuildPlan` graph reconstruction

* Minimise I/O by replacing BuildProfile's PathMap with a ManifestMap

This commit reduces the number of times that we need to read each
dependency's manifest from file in order to reduce the amount of I/O
performed. Each dependency's manifest should now only be loaded once
during either `graph_to_manifest_map` or `fetch_graph`.

This required implementing `Clone` for `ManifestFile` to allow for
constructing the build plan in a manner that does not require consuming
the manifest file. `ManifestFile::clone` should only ever be called for
the root node.

As a side-effect, the `build` and `check` commands no longer require
loading each package's manifest as they are already provided by the
build plan. As a result both commands no longer require the sway git
tag to be passed through.

* Ensure only one edge between a package and each dependency

* Ensure lock is updated if path child dependencies change

* Address clippy nits

* Address formatting nits

* Improve accuracy of dependency removal by invalidating by edges

Previously, all dependency nodes were removed in the case that they were
invalid for any one of their parents.

This commit changes the behaviour to not remove the node, but rather
each invalid dependency *edge* instead. This means that if a node is
valid for one parent but not another, only the edge to the other parent
is removed and not the whole node itself. The node is only removed in
the case that it is invalid for all of its parents.

This is necessary for handling the case where a dependency no longer
exists in one parent's `[dependencies]` but still exists in another. The
most common case for this is `std`, as many nodes may depend on `std`
however we don't want to remove `std` and have to re-fetch it just
because one of the nodes removed their depenency on it.

* Avoid visiting a node's dependencies multiple times during validation

* Improve name and docs for BuildPlan constructors

* Remove unnecessary path root validation from Lock::to_graph.

This is checked during the `validate_graph` phase.

Also improves a bunch of comments.

Co-authored-by: kayagokalp <kaya.gokalp@fuel.sh>
Co-authored-by: Chris O'Brien <57543709+eureka-cpu@users.noreply.github.com>
Co-authored-by: Kaya Gökalp <kayagokalp@sabanciuniv.edu>