Add a new calling context attribute to restrict functions to particular contexts.
Adds support for an attribute (i.e. `#[context(internal_only)]`) to
annotate the supported calling contexts for particular functions.

This annotation is then be used by the compiler to check and restrict
the calling of these functions to a compatible calling context.

The main goal is to be able to get compile-time safety when working with
functions that should only be called from an internal (contract)
context.

A real-world example of this is the [msg_amount() stdlib
function](https://github.com/FuelLabs/sway-lib-std/blob/6a8f5bf588df5d0679fb834f05c900f2e54de426/src/context.sw#L33-L38)
which only returns a proper value when used in an internal context. In
an external context is returns a well-defined but not semantically
representative value, specifically 0, as explained by @adlerjohn in
[FueLabs/sway#963](https://github.com/FuelLabs/sway/issues/963).

Closes https://github.com/FuelLabs/sway/issues/963.