Collect tokens into a `HashMap` instead of a `Vec` in the language server. (#2121)
This introduces the following type for collecting tokens during parsing.

```
pub type TokenMap = HashMap<(Ident, Span), TokenType>;

#[derive(Debug, Clone)]
pub struct TokenType {
    pub parsed: AstToken,
    pub typed: Option<TypedAstToken>,
}
```

First, we traverse the untyped parse tree and collect the AstToken into parsed. Then we traverse the typed_parse_tree and optionally collect a TypedAstToken if one was found. After both traversal stages, we end up with a HashMap that contains all of the tokens found in the sway file.

This PR replaces the current Token module with the above data structure.