Parse tokens from all dependencies instead of single files in the language server (#2346)
This PR now traverses the submodules in a project and collects those tokens into the TokenMap.

I've moved a lot of the logic that was previously in document.rs into session.rs. Also ended moving / simplifying a bit of code so it's easier to read.

Finally, I've changed our TokenMap type to use a DashMap instead of a HashMap.