Fix concurrency bug in ConcurrentSlab.replace.
So what happens here is that we get the following panic during
`ConcurrentSlab.replace` execution:

```
thread 'main' panicked at 'rwlock read lock would result in deadlock',
/rustc/ddcbba036aee08f0709f98a92a342a278eae5c05/library/std/src/sys/unix/locks/pthread_rwlock.rs:72:13
```

This happens due to `operator !=`, we call `PartialEq` and eventually
end up calling `ConcurrentSlab.get`, which fails due to existing
read/write lock already on the stack.

The fix is to make sure to get a read lock only, do the comparison,
and then get a read/write lock just for the write.

Closes https://github.com/FuelLabs/sway/issues/2341.