Collect more tokens in the language server and store definitions (#2409)
Most of the changes in the PR result from renaming TokenType to Token. I did this as we are now storing the type definition of the token during the collection phase. This can either be a TypeId or the Ident of the declaration token.

The rest is just collecting more tokens that I noticed were missing when working through the e2e programs. There are still more tokens that need collecting but happy to address them in separate PR's.