Fix a bug in the IR parsing.
When resolving calls after recreating the IR from parsed input as a
second pass the old code would inject a NOP where the CALL should go and
replace it with a resolved CALL.

This was a problem when the call value was used as an argument to another
call, as the NOP was taken rather than the CALL.  Attempting to replace
the NOP everywhere was very tricky (I couldn't actually manage it in the
end).

A better (though still a bit hacky) approach now is to inject a CALL
from the start, but to a dummy function and replace that function within
the CALL in the resolving pass.  This way any references to the call
value are valid from the start.