Implement a runnables custom method and collect typed tokens (#2441)
This PR enables the language server to collect and use typed tokens. As a result, go to definition works for tokens that have an associated type definition (roughly 50% of tokens).

This also allows us to get the main_function for script and predicate files, allowing us to send the code position on the main fn back to the client. Code lens is implemented on the client side for now which allows us to have a play button show up above the main fn to execute forc run.

Currently, we are only able to use the typed tokens, and send information about the main fn location if the program is able to successfully compile. This should be fixed once #1674 is addressed.