Speed improvements (and a buffix) for simplify-cfg.
- New algorithm for `remove_dead_blocks`
- Cached predecessor counts for `merge_blocks`

Bugfix: BlockContent::successors() was incorrect. It was
returning successors for all blocks in the function.

With this change, we now can compile the `stdlib/vec` test in about
35s (on my i7-7th-gen/32GB RAM laptop).

Prio to this change, it was [previously taking](https://github.com/FuelLabs/sway/pull/2548#issue-1340432847) about 25m.

Testing: The `stdlib/vec` test's bytecode with this change is identical
to the one without this change. The existing regression tests
for simplify-cfg pass.

I've also enabled this pass by default now.