Collect SymbolKind info and generate correct semantic tokens (#2569)
This PR introduces a new SymbolKind enum which is stored in the Token type. We are populating the kind of each token during the parse-tree traversal stage. This makes it a lot cleaner when going from our Token type to the various lsp_types enums such as CompletionItemKind, SemanticTokenType etc..

We are now providing accurate semantic token information to the client which is resulting in much nicer syntax highlighting for sway files.