Fix warning regression when handling break and continue expressions.
This fixes a recently introduced regression when checking for unused
expression values inside blocks.

The following code:

```rust
script;

fn main() {
    while true {
        if true {
            break;
        }
    }
}
```

Would emit the following warning:

```
This returns a value of type unknown, which is not assigned to anything
and is ignored.
```

The fix is pretty simple and just checks to make sure we are not
handling either a break or return expresisons before issuing the
warning.

Closes https://github.com/FuelLabs/sway/issues/2638.