Collect missing tokens in the language server (#2620)
This PR collects most of the remaining tokens in the language server.

I wrote a script that saved out the ASTs for the sway_ast::Module, ParseProgram and TypedProgram stages for each of the sway examples and all of the e2e/language/should pass examples. This made it quick to inspect what extra span information needed to be carried through in the compiler.

There are still a few tokens not being collected but they don't involve trivial changes to the compiler so I'll address them individually in separate PRs.