Update the tests for non-inlined function calls.
Most of the changes are due to the banning of non-copy-type constants.
Previously we could just whip up an empty struct from thin air and
populate it, which _actually_ meant creating it on the stack and leaking
it.  This is no longer possible with function calls and balanced
call-frames on the stack.