Change `asm_lang` a bit to support the new back end pipeline.
We used to go 'virtual' -> 'realised' -> 'allocated'.  Now it's changed
to 'virtual' -> 'allocated' -> 'realised'.

This means `OrganizationalOp` is parameterised into `ControlFlowOp` so
we can have non-allocated and allocated 'abstract' (non-realised)
opcodes.

And we now have `AllocatedAbstractOp` which is the result of register
allocation and is fed to address realisation.

This has been hacked together to take what we used to have and make it
work with what we now need.  It is poorly designed and desperately needs
to be refactored into something with better types and less redundancy.