Forward attributes from `ParseProgram` to `TyProgram`
This is useful for both upcoming `forc doc` (`#[doc]`) and `forc test`
(`#[test]`) behaviour.

`forc test` requires this for in-language unit tests as it needs access
to the attributes of the function declaration during dead code analysis.
This is so that we can add the test functions as entry points and avoid
producing dead function warnings for all `#[test]` functions.

@eureka-cpu also mentioned that this would be useful for `#[doc]`
support, though I don't quite remember the details.