LSP synchronize workspace (#2919)
Previously, the language server was falling out of sync with the users
current workspace edits. For example, incorrect syntax highlighting can
be seen in the below image if the user creates a new line above the
function. The syntax then takes on the previous highlighting and all
other attributes like definition locations, warnings/error locations
etc..

<img width="700" alt="Screen Shot 2022-10-12 at 12 39 47 pm"
src="https://user-images.githubusercontent.com/1289413/195230126-9942e54a-5f61-4a3f-9efa-cf1d21653615.png">

This was happening because we load the project from disk and then get
ASTs to work with from the compiler from these files. As the users
workspace does not match the files saved on disk, then incorrect `Span`s
are in the AST. To get around this, we are cloning the users workspace
into a temporary directory on initialization of the server. On each
keystroke or edit, we update the temp files and save to disk so that we
can get an accurate AST back from the compiler for the language server
to work with. The temp directory is synchronized with the users
workspace on each did_save event.

Finally, we spawn a tokio thread and use the `notify` crate to look for
changes to the projects `Forc.toml` file. We then convert any
dependencies that have a relative path to an absolute path.

closes #2847

I've added error handling now as #2968 was merged before this one was
approved.