Speedup `remove_redudant_moves()` by using a `HashSet` instead of a `BTreeSet` (#3099)
As we're not actually looping over the set of used registers, it can
just be a `HashSet`. I did not try to optimized this further.

Tested with [the `amm` repo](https://github.com/sway-libs/amm) (without
https://github.com/FuelLabs/sway/pull/3096):
Previously:
```console
❯ forc b --time-phases --silent
...
  Time elapsed to compile to ast: 4.8512885s
  Time elapsed to generate JSON ABI program: 50.875µs
  Time elapsed to compile ast to asm: 53.342876291s
  Time elapsed to compile asm to bytecode: 27.719744166s
```
With this change:
```console
❯ forc b --time-phases --silent
...
  Time elapsed to compile to ast: 4.629979583s
  Time elapsed to generate JSON ABI program: 53.5µs
  Time elapsed to compile ast to asm: 6.345263583s
  Time elapsed to compile asm to bytecode: 28.130886833s
```
Specifically looking at the second to last step which is ~8.4X faster.