LSP: Show formatted documentation on hover (#3084)
This PR renders adds MarkDown documentation hover support. It works by
using the doc attributes that can be found for the declaration of the
currently hovered token in the editor.

We can also have blocks of ```sway code in the docs and it will render
those sections out with correct sway syntax highlighting :)

I've borrowed and modified the `markdown.rs` from `rust-analyzer` to get
this to work. For example, consider we have the following doc comments
above a type in our program.

```
/// Struct holding:
///
/// 1. A `value` of type `NumberOrString`
/// 2. An `address` of type `u64`
/// 2. A `name` of type `str[11]`
///
/// Lets look at some formatted sway code below.
///
/// ```sway
/// let value = NumberOrString::Number(20);
/// let address = 0; 
/// let name = "my_contract";
/// ```
/// 
/// Yay, we now have rich markdown formatting of doc comments!
struct Data {
    /// The `value` field in `Data`
    value: NumberOrString,
    /// The `address` field in `Data`
    address: u64,
    /// The `name` field in `Data`
    name: str[11],
}
```

If we hover over the `Data` type when it's used elsewhere in the
program, we get the following hover box appearing.

<img width="632" alt="Screen Shot 2022-10-20 at 3 14 38 pm"
src="https://user-images.githubusercontent.com/1289413/196855038-2961f237-a9d0-4e72-b502-fa42d522fbc4.png">

closes #2849