Avoids inline when #[inline(never)] is used.
inline_function_calls' inline_heuristic now checks inline metadata and
if inline metadata is set to Inline::Never it returns false and the
function is not inlined.

When inline metadata is set to Inline::Always it will still follow the
default behaviour, I left as TODO: check if inlining of function is
possible.

ref #2944