Use a call-graph in the inliner (#3186)
This reduces the time spent in the inliner, in compiling the [`amm`
library ](https://github.com/sway-libs/amm) from ~120s to ~5s.

Code sizes before/after on
[sway-applications](https://github.com/FuelLabs/sway-applications):
-----------------------------------
benchmark | before | after
----------------|----------|--------
dao-voting | 20568 | 20496
escrow | 72740 | 72668
fundraiser | 31344 | 31440
multisig-wallet | 6592 | 7000
NFT | 23200 | 23200
oracle | 1924 | 1924

While this change doesn't touch the inlining heuristics, but only the
order in which inlining of functions is undertaken, we have a heuristic
that checks the size of the function being inlined. **That** depends on
the order in which we inline, and hence the final set of inlining
decisions aren't identical (though largely similar). Overall, I think
the compile time benefits outweigh any differences in the inlining
decisions.