Fixed "Cannot infer type" error to point to call site span.
`CompileError::UnableToInferGeneric` was using as span the span
of the type_id. This causes the error to point to where the type is
declared instead of to where it should be annotated.

The solution was to store a call site span in `CollectTypesMetadataContext`
and pass it to `TypeInfo::UnknownGeneric` so when we convert it to the
actual error we can use the call site span which points to the place
that is missing a type annotation.

Closes #3192