Add intrinsics for operating on raw_ptrs (#3157)
This PR:
- Adds `__ptr_add<T>(ptr: raw_ptr, count: u64)` (which basically does
`ptr + __size_of::<T>() * count`) and its sub counterpart to be used in
pointer offset calculations.
- Makes `std::alloc::*` and `raw_ptr::*` have type arguments and work in
increments of `size_of::<T>()` instead of bytes.
- Removes test `fixing_generic_type` which was broken instead of
maintaining it. This test was created to reproduce an issue I was having
a few months ago. Fixing this allowed other things to be built so there
are other tests that touch this case.