Fix CFG node uniqueness.
The CFG now keeps track of previously processed declarations and makes
sure to re-use the previously added graph node, instead
of creating duplicate nodes for the same declarations.

We do this by introducing a new IdentUnique type, which is pretty much
the same as an Ident, but hashes with the full span, as opposed to an
Ident which just uses the name as the hash, hence not fully ideal
to use as the key for the CFG namespace, since it won't work
correctly in multi-file cases.

Closes https://github.com/FuelLabs/sway/issues/3093.