Some micro-optimizations to `U128` and `U256` (#3253)
Once we have a fully fledged instruction combining optimization pass,
these types of optimizations won't be needed anymore (for the most
part). For now, this allows us to re-enable the `u128_root_test` that
was disabled in #3181.
* Switch division by 2 to a shift left by 1
* Switch to using the lower part of a `U128` in situations where the
upper part is obviously not impacted.

This cuts down the gas cost for `sqrt` by around 30-40% and the
`u128_root_test` now consumes ~77M gas. Previously, the test was running
out of gas (our default limit is 100M).

I also made the same optimization in `U256` but notice that the
`u256_div_test` was disabled (doesn't have a `test.toml`!) so I enabled.
The test was actually running out of gas before my change but now it
just barely passes with like 98M gas, which is dangerous.

Finally, I replaced the method of multiplication in both `U128` and
`U256` and enabled two previously disabled tests:
* `u256_mul_test` which used to run out of gas but now takes ~21M (so
that's a least 5 X improvement).
* `u256_ops_test` which runs fine.

In short, div and mul for both `U128` and `U256` are now much cheaper
and so is `sqrt` for `U128` (`U256` doesn't have a `sqrt` yet).

Closes #3256 
Closes #3245

Co-authored-by: João Matos <joao@tritao.eu>