Fixes method Eq not found when used with where clause.
Solution was to change how `TypeInfo::UnknownGeneric` are compared in
`are_equal_minus_dynamic_types`. It now compares the span of the name
which is unique in the source code. If the span is equal then it assumes
it is safe to consider the TypeIds refer to the same generic parameter
and that using the methods of one another is also safe.

Closes #3324