Refactor E2E CLI runner.
Now output handling is done in an uniform manner for all kinds of tests,
which makes the code easier to reason, and makes it possible to
implement better developer exeperience.

Now the output is hidden by default for all tests, and only shown up
when a particular test is failing (indented so its easier to parse).

Failure handling is also more robust, with all tests being executed by
default, and failures being reported as they happen and in the summary
report at end of execution. There are still a couple panics here and
there for more esoteric errors, but those can be easily be transformed
to errors in the future as need arises.

Previous behavior of showing all output can be obtained by passing the
`--verbose` flag to the runner.

The only exception are the IR tests which still panic. However their
summary report was updated to match the improvements from above.

Closes https://github.com/FuelLabs/sway/issues/3060.