Refactor E2E CLI runner. (#3294)
Now output handling is done in an uniform manner for all kinds of tests
and makes it possible to implement better developer experience.

Now the output is hidden by default for all tests, and only shown up
when a particular test is failing (indented so its easier to parse).

Failure handling is also more robust, with all tests being executed by
default, and failures being reported as they happen and in the summary
report at end of execution. There are still a couple panics here and
there for more esoteric errors, but those can be easily be transformed
to errors in the future as need arises.

Previous behavior of showing all output can be obtained by passing the
`--verbose` flag to the runner.

The only exception are the IR tests which still panic. However their
summary report was updated to match the improvements from above.

Closes https://github.com/FuelLabs/sway/issues/3060.

Example of new output with failed test report because `fuel-core` is not
running:

![Screenshot from 2022-11-09
17-53-05](https://user-images.githubusercontent.com/602268/200904172-c9327b78-b557-49b4-9e4f-d78c3d5a2d7b.png)

Co-authored-by: Mohammad Fawaz <mohammadfawaz89@gmail.com>