Disallows while loops in predicate dependencies.
Current solution was not throwing the error when the predicate
uses a library with loops.
We still get the `The JI opcode cannot jump backwards in a predicate.`
but it is not very useful because it has no span available.

We could check for while loops in other modules too but `type_check`
checks all declarations so even functions that are not used would
thrown an error if they contain loops. Ideally we want to throw an
error only if the predicate uses a function with a while loop.

The solution was to perform the while in predicate check on
`compile_while_loop` as at this time only used functions are compiled.