Add `#[payable]` annotation to ABI methods
The lack of this annotation implies the method is nonpayable.
When calling an ABI method that is nonpayable, the compiler must emit an error
if the amount of coins to forward is not guaranteed to be zero.

Currently, the analysis of non-zero coins is very simple and only checks
if the special `coins` contract call parameter is the zero u64 literal
directly or can be reached through a chain of variable/constant definitions.

The compiler also emits an error if the method signature and
the method implementation have different `#[payable]` annotations.

close #1608