Do not progress with IR generation if `deterministically_aborts` (#3460)
Fixes #3303 

The existing check for `deterministically_aborts` assumes that function
applications are inlined. This shouldn't be so for the newly added call
to `determinisitically_aborts`. So I've added a new parameter to the
`DeterministicallyAborts` trait to switch considering "calls are
inlined" or not semantics. I'm not sure if this is the best way, so I'm
open to suggestions.