Fixes compiler picking wrong trait implementation for tuples.
Tuples with same fields types but different lengths were consider equal
by are_equal_minus_dynamic_types and an error was thrown later on when
the number of fields didn't match.

The solution was to fix are_equal_minus_dynamic_types to check tuple
length before doing a zip of the field iterators.

Closes #3485