Fixes error when dependency variable is used as function parameter.
When a dependency variable was passed into a function as first parameter
such as `ContractId::from(contract_c::CONTRACT_ID);` an error was thrown
by `type_check_method_application` while it tried to resolve the
variable as a local one.

The solution was to only resolve the variable if the function is
expecting a mutable parameter.

Closes #3406