Fix DCA warnings for log intrinsic.
This adds special handling for the log intrinsic, consuming the fields
of all structs used inside the inner expressions.

It also does the same for functions external to the current analysis,
which could lead to some false negatives when it comes to the unused
fields warning, but it's the best we can do for now without considerable
effort refactoring the DCA engine to work inter-proceduraly/module.

Closes https://github.com/FuelLabs/sway/issues/3012.