Fix DCA warnings for impl trait methods.
We now connect the edge of methods referenced in function call
expressions to the edge of the corresponding node of the implemented
trait method.

For this FunctionDeclaration was extended to keep track of the
implementing type where it is declared.

Also previously we were not correctly connecting edges in
add_edge_from_entry due to the function using self.entry_points which is
only computed as the last stage of the DCA process. Fix this by
deferring these edge node indices and only making the connection after
they are computed.

Aditionally now we use an `IdentUnique` to keep track of functions in
the CFG namespace which helps disambiguate functions shadowing each
other.

Closes https://github.com/FuelLabs/sway/issues/3067.