formatter: fix handling of statements ending with `;` in `if else` blocks (#3557)
unblocks #3556
closes #3555 

~The situation occurring in the issue seems to happen when the formatter
tries to (incorrectly) assume that an if expression can be in-lined when
there's a newline character inside. During this check, the line kind
(multi-line vs inline) is decided solely depending on the width of the
line, ignoring the existence of a `\n` character.~

~This PR introduces something like a 'catch-all' during the above check
to default to `LineKind::Multiline` if a `\n` is found during the
look-ahead formatting done by the temporary formatter, which forces the
formatter to use multi-line style, so that at the very least we don't
end up with a broken inline format as seen in the issue.~

~This catch-all might unblock the comment formatting PR above as well.~

The above was my misunderstanding - the mistake was further upstream
instead within statement formatting. The inline style was not being
handled for let statements.