Re-order the blocks in while loops to help the register allocator.
Unfortunately the register allocator depends on value uses to always
follow value definitions in the list of instructions within a function.

Until now while loops were placing the body beyond the end of loop block
which violated this requirement.

This change introduces a special 'break block' to allow the while loop
blocks to be in top-to-bottom order again.  This break block is then
mostly optimised away with `simplify-cfg`.