Add a `cargo package` CI job to check pkgs are publish-friendly
`cargo package` assembles a local package into a distributable tarball.
It can be thought of as doing a dry-run for the publish command, however
unlike `cargo publish --dry-run`, the `cargo package` command is
compatible with running at the workspace level (as introduced in this
PR) which makes it a little easier to keep up to date for all packages
in our workspace.

The main benefit of this job is to catch errors caused by the sandboxed
build, e.g. attempting to statically link to assets or other projects
outside of the project's directory.

This job *might* take an excessive amount of time due to building each
package one-at-a-time in the sandboxed environment. I'm not sure if the
Rust cache that we use accounts for caching projects built via `cargo
package`. We'll have to keep an eye on the times and see if it's worth
it!